requirements.txt
fastapi
uvicorn
pandas
requests
pyjwt
